+++ lib/dotenv_configuration_provider.dart
@@ -1,7 +1,12 @@
 import 'package:configurable/configuration_provider.dart';
 import 'package:configurable/dotenv/dotenv.dart' show env, load;
 
+/// Dotenv Configuration Provider
+///
+/// @author sombochea
+/// @since 1.0.0
 class DotenvConfigurationProvider implements ConfigurationProvider {
+  /// default constructor to load the platform environment
   DotenvConfigurationProvider() {
     load();
   }
@@ -15,6 +20,7 @@ class DotenvConfigurationProvider implements ConfigurationProvider {
   String? getOrNull(String key, {String? defaultValue}) {
     var value = env[key];
 
+    /// if value is null, then return default value
     if (value == null) {
       return defaultValue;
     }