+++ test/config_test.dart
@@ -1,5 +1,5 @@
 import 'package:configurable/configurable.dart' show getConfigOrNull, hasConfigkey;
-import 'package:configurable/dotenv_configuration_provider.dart';
+import 'package:configurable/simple_configuration_provider.dart';
 import 'package:configurable/system_config.dart';
 import 'package:test/test.dart';
 
@@ -7,20 +7,11 @@ void main() {
   test('get system config by key', () {
     var key = 'app.name';
     var value = 'CUBETIQ';
+    
+    // set simple provider
+    SystemConfig.setProvider(SimpleConfigurationProvider());
+
     var result = SystemConfig.getOrNull(key, defaultValue: value);
-
-    expect(value, equals(result));
-    expect(value, equals(SystemConfig.getOrNull(key)));
-  });
-
-  test('get system config by key with dotenv provider', () {
-    var key = 'app.name';
-    var value = 'CUBETIQ Solution';
-
-    // set dotenv provider
-    SystemConfig.setProvider(DotenvConfigurationProvider());
-
-    var result = SystemConfig.getOrNull(key);
 
     expect(value, equals(result));
     expect(value, equals(SystemConfig.getOrNull(key)));
@@ -29,6 +20,9 @@ void main() {
   test('get config by key with function', () {
     var key = 'app.name';
     var value = 'CUBETIQ Solution';
+
+    // set simple provider
+    SystemConfig.setProvider(SimpleConfigurationProvider());
 
     var result = getConfigOrNull(key, defaultValue: value);