+++ CHANGELOG.md
@@ -10,3 +10,9 @@
 
 ## 1.0.2
 - Add support null-safety
+
+## 1.0.3
+- Add mutable configuration provider
+- Split tests file
+- Fixed and nullable functions
+- Able to set or remove config from system config